test: split up huge test module
